+++ src/main/java/com/programmers/voucher/repository/MemoryVoucherRepository.java
@@ -1,13 +1,10 @@
 package com.programmers.voucher.repository;
 
 import com.programmers.voucher.domain.Voucher;
-import com.programmers.voucher.dto.VoucherRequestDto;
 import com.programmers.voucher.dto.VoucherResponseDto;
 import org.springframework.stereotype.Repository;
-
 import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.stream.Collectors;
 
 @Repository
 public class MemoryVoucherRepository implements VoucherRepository {
@@ -15,9 +12,9 @@ public class MemoryVoucherRepository implements VoucherRepository {
     private final static Map<UUID, Voucher> storage = new ConcurrentHashMap<>();
 
     @Override
-    public Voucher save(Voucher voucher) {
+    public UUID save(Voucher voucher) {
         storage.put(voucher.getVoucherId(), voucher);
-        return voucher;
+        return voucher.getVoucherId();
     }
 
     @Override
