+++ src/main/java/com/programmers/voucher/repository/VoucherRepository.java
@@ -3,7 +3,6 @@ package com.programmers.voucher.repository;
 import com.programmers.voucher.domain.Voucher;
 import com.programmers.voucher.dto.VoucherResponseDto;
 import org.springframework.stereotype.Repository;
-
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
@@ -11,7 +10,7 @@ import java.util.UUID;
 @Repository
 public interface VoucherRepository {
 
-    Voucher save(Voucher voucher);
+    UUID save(Voucher voucher);
 
     List<VoucherResponseDto> findAll();
