+++ src/main/java/com/programmers/voucher/service/VoucherService.java
@@ -1,36 +1,13 @@
 package com.programmers.voucher.service;
 
-import com.programmers.console.util.VoucherPolicy;
-import com.programmers.voucher.domain.Voucher;
 import com.programmers.voucher.dto.VoucherRequestDto;
 import com.programmers.voucher.dto.VoucherResponseDto;
-import com.programmers.voucher.repository.VoucherRepository;
-import org.springframework.stereotype.Service;
-
-import java.time.LocalDate;
 import java.util.List;
 import java.util.UUID;
 
-@Service
-public class VoucherService {
+public interface VoucherService {
 
-    private final VoucherRepository voucherRepository;
+    UUID create(VoucherRequestDto voucherRequestDto);
 
-    public VoucherService(VoucherRepository voucherRepository) {
-        this.voucherRepository = voucherRepository;
-    }
-
-    public Voucher create(VoucherRequestDto voucherRequestDto) {
-        Voucher voucher = toEntity(UUID.randomUUID(), voucherRequestDto);
-        return voucherRepository.save(voucher);
-    }
-
-    public List<VoucherResponseDto> findVouchers() {
-        return voucherRepository.findAll();
-    }
-
-    private Voucher toEntity(UUID voucherId, VoucherRequestDto voucherRequestDto) {
-        return VoucherPolicy.findByCommand(voucherRequestDto.getVoucherType())
-                .constructor(voucherId, voucherRequestDto.getDiscountAmount(), LocalDate.now());
-    }
+    List<VoucherResponseDto> findVouchers();
 }
